fix(20201021211638_create-locations.js): zip column in locations table should be a string
zip codes can contain up to 4 '0's e.g '00001' is a valid zip code. however if we insert this as a
integer in the databases it will only insert as the '1', due to this zip codes should be stored as a
string. an addtional reason is computation is never done on zip codes as they are created for humans
to read easily.